Updated file structure, added subdirectories for panels. RunProgram and TraceSession hooks.
These changes were mostly made to further keep application logic away from Main.jsx.

On run, Main.jsx calls useRunProgram.js now, which is a hook that calls the backend to execute the program.

Similarly, on trace Main.jsx calls useTraceSession that calls the backend to execute the trace logic (not fully implemented, still kinda broken. we need to update the routes)

I also refactored the css to use modules. Now each panel has its own styles, so its easier to keep track of what's doing what.
