+++ weblcc/src/components/LeftPanel/Editor/EditorPanel.module.css
@@ -1,0 +1,67 @@
+.panel {
+  background: var(--bg1);
+  border: 1px solid var(--border2);
+  border-radius: 10px;
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  min-height: 0;
+  animation: fadeUp 0.3s ease;
+}
+
+.header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 10px 12px;
+  font-size: 0.8rem;
+  color: var(--text2);
+  border-bottom: 1px solid var(--border2);
+  text-transform: uppercase;
+  letter-spacing: 0.18em;
+}
+
+.body {
+  flex: 1;
+  overflow: auto;
+  display: flex;
+  flex-direction: column;
+}
+
+.textarea {
+  flex: 1;
+  width: 100%;
+  resize: none;
+  border: none;
+  background: var(--bg3);
+  color: var(--text);
+  padding: 12px;
+  font-family: "JetBrains Mono", monospace;
+  font-size: 0.85rem;
+  outline: none;
+}
+
+.codeDisplay {
+  padding: 12px;
+  font-family: "JetBrains Mono", monospace;
+  font-size: 0.85rem;
+}
+
+.codeLine {
+  display: grid;
+  grid-template-columns: 40px 1fr;
+  gap: 8px;
+  padding: 2px 0;
+}
+
+.codeLineNumber {
+  color: var(--text3);
+  text-align: right;
+}
+
+.codeLineActive {
+  composes: codeLine;
+  background: var(--gold-dim);
+  border-left: 2px solid var(--gold);
+  padding-left: 8px;
+}
