+++ weblcc/src/pages/Main.jsx
@@ -1,8 +1,18 @@
 import { useMemo, useState } from 'react';
-import EditorPanel from '../components/EditorPanel.jsx';
-import OutputPanel from '../components/OutputPanel.jsx';
-import RightPanel from '../components/RightPanel.jsx';
-import Topbar from '../components/Topbar.jsx';
+
+import { DEFAULT_FLAGS, DEFAULT_REGISTERS, DEFAULT_STACK } from '../constants/cpuDefaults.js';
+import { useRunProgram } from '../hooks/useRunProgram.js';
+import { useTraceSession } from '../hooks/useTraceSession.js';
+
+import EditorPanel from '../components/LeftPanel/Editor/EditorPanel.jsx';
+import OutputPanel from '../components/LeftPanel/Terminal/OutputPanel.jsx';
+import Topbar from '../components/Header/Topbar.jsx';
+import FlagsPanel from '../components/RightPanel/Flags/FlagsPanel.jsx';
+import MemoryPanel from '../components/RightPanel/Memory/MemoryPanel.jsx';
+import RegistersPanel from '../components/RightPanel/Registers/RegistersPanel.jsx';
+import StackPanel from '../components/RightPanel/Stack/StackPanel.jsx';
+
+import styles from './Main.module.css';
 
 const starterProgram = `; ex0301.a
           ld r0, x
@@ -21,119 +31,31 @@ function normalizeLines(source) {
 
 export default function Main() {
   const [source, setSource] = useState(starterProgram);
-  const [isRunning, setIsRunning] = useState(false);
-  const [output, setOutput] = useState('');
-  const [traceSteps, setTraceSteps] = useState([]);
-  const [currentStep, setCurrentStep] = useState(0);
-  const [isTracing, setIsTracing] = useState(false);
   const [leftTab, setLeftTab] = useState('output');
-  const [rightTab, setRightTab] = useState('registers');
+
+  const { isRunning, output, setOutput, handleRun } = useRunProgram();
+  const {
+    traceSteps,
+    currentStep,
+    setCurrentStep,
+    isTracing,
+    stepCount,
+    activeStep,
+    handleTrace,
+    handlePrev,
+    handleNext,
+    handleReset,
+    canStepBack,
+    canStepForward
+  } = useTraceSession();
 
   const lines = useMemo(() => normalizeLines(source), [source]);
-  const activeStep = traceSteps[currentStep];
   const activeLine = activeStep ? activeStep.lineNumber - 1 : -1;
-  const stepCount = traceSteps.length;
 
-  const registers = activeStep?.registers || [
-    { name: 'r0', value: '0x0000' },
-    { name: 'r1', value: '0x0000' },
-    { name: 'r2', value: '0x0000' },
-    { name: 'r3', value: '0x0000' },
-    { name: 'r4', value: '0x0000' },
-    { name: 'r5', value: '0x0000' },
-    { name: 'r6', value: '0x4000' },
-    { name: 'r7', value: '0x0000' }
-  ];
-
-  const flags = activeStep?.flags || { N: 0, Z: 0, P: 0, C: 0, V: 0 };
-
-  const stack = activeStep?.stack || [
-    { addr: '0x4000', value: '0x0000' },
-    { addr: '0x3fff', value: '0x0000' },
-    { addr: '0x3ffe', value: '0x0000' }
-  ];
-
+  const registers = activeStep?.registers || DEFAULT_REGISTERS;
+  const flags = activeStep?.flags || DEFAULT_FLAGS;
+  const stack = activeStep?.stack || DEFAULT_STACK;
   const eventLabel = activeStep?.action || 'No trace data loaded.';
-
-  const memoryRows = useMemo(() => {
-    const rows = [];
-    for (let row = 0; row < 6; row += 1) {
-      const addr = `0x${(0x4000 + row * 4).toString(16).padStart(4, '0')}`;
-      const cells = [];
-      for (let col = 0; col < 4; col += 1) {
-        const index = row * 4 + col;
-        const value = stack[index]?.value || '0x0000';
-        cells.push(value);
-      }
-      rows.push({ addr, cells });
-    }
-    return rows;
-  }, [stack]);
-
-  async function handleRun() {
-    setIsRunning(true);
-    setOutput('');
-
-    try {
-      const response = await fetch('/api/run', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ source })
-      });
-
-      const data = await response.json();
-      if (!response.ok) {
-        throw new Error(data.error || 'Failed to run program');
-      }
-
-      const combined = [data.stdout, data.stderr].filter(Boolean).join('\n');
-      setOutput(combined || 'Program finished with no output.');
-    } catch (error) {
-      setOutput(String(error));
-    } finally {
-      setIsRunning(false);
-    }
-  }
-
-  async function handleTrace() {
-    setOutput('');
-    setIsTracing(true);
-    setCurrentStep(0);
-
-    try {
-      const response = await fetch('/api/trace', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ source })
-      });
-
-      const data = await response.json();
-      if (!response.ok) {
-        throw new Error(data.error || 'Failed to trace program');
-      }
-
-      setTraceSteps(data.steps || []);
-      setLeftTab('trace');
-    } catch (error) {
-      setOutput(String(error));
-      setTraceSteps([]);
-      setIsTracing(false);
-    }
-  }
-
-  function handlePrev() {
-    setCurrentStep((prev) => Math.max(prev - 1, 0));
-  }
-
-  function handleNext() {
-    setCurrentStep((prev) => Math.min(prev + 1, stepCount - 1));
-  }
-
-  function handleReset() {
-    setTraceSteps([]);
-    setCurrentStep(0);
-    setIsTracing(false);
-  }
 
   function handleDownload() {
     const blob = new Blob([source], { type: 'text/plain' });
@@ -148,47 +70,45 @@ export default function Main() {
   }
 
   return (
-    <div className="app-shell">
+    <div className={styles.appShell}>
       <Topbar
         isRunning={isRunning}
         stepCount={stepCount}
-        onRun={handleRun}
-        onTrace={handleTrace}
+        onRun={() => handleRun(source)}
+        onTrace={() => handleTrace(source, () => setLeftTab('trace'), setOutput)}
         onPrev={handlePrev}
         onNext={handleNext}
         onReset={handleReset}
         onDownload={handleDownload}
-        canStepBack={stepCount > 0 && currentStep > 0}
-        canStepForward={stepCount > 0 && currentStep < stepCount - 1}
+        canStepBack={canStepBack}
+        canStepForward={canStepForward}
       />
 
-      <div className="layout">
-        <EditorPanel
-          isTracing={isTracing}
-          lines={lines}
-          activeLine={activeLine}
-          source={source}
-          onSourceChange={setSource}
-        />
+      <div className={styles.layout}>
+        <div className={styles.left}>
+          <EditorPanel
+            isTracing={isTracing}
+            lines={lines}
+            activeLine={activeLine}
+            source={source}
+            onSourceChange={setSource}
+          />
+          <OutputPanel
+            activeTab={leftTab}
+            onTabChange={setLeftTab}
+            output={output}
+            traceSteps={traceSteps}
+            currentStep={currentStep}
+            onSelectStep={setCurrentStep}
+          />
+        </div>
 
-        <OutputPanel
-          activeTab={leftTab}
-          onTabChange={setLeftTab}
-          output={output}
-          traceSteps={traceSteps}
-          currentStep={currentStep}
-          onSelectStep={setCurrentStep}
-        />
-
-        <RightPanel
-          activeTab={rightTab}
-          onTabChange={setRightTab}
-          registers={registers}
-          stack={stack}
-          flags={flags}
-          memoryRows={memoryRows}
-          eventLabel={eventLabel}
-        />
+        <div className={styles.panelRight}>
+          <RegistersPanel registers={registers} />
+          <FlagsPanel flags={flags} />
+          <StackPanel stack={stack} eventLabel={eventLabel} />
+          <MemoryPanel stack={stack} />
+        </div>
       </div>
     </div>
   );
